프로젝트 패키지 구조 layer > 모듈 기준으로 변경
