+++ src/main/java/com/teamProject/syusyu/controller/CommentController.java
@@ -5,7 +5,6 @@ import com.teamProject.syusyu.service.CommentService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.http.HttpSession;
