+++ src/main/java/com/teamProject/syusyu/domain/PageHandler.java
@@ -1,10 +1,5 @@
 package com.teamProject.syusyu.domain;
 
-import org.springframework.web.util.UriComponentsBuilder;
-
-import java.net.SocketTimeoutException;
-
-
 
 public class PageHandler {
 //    private int page; // 현재 페이지
