+++ src/main/java/com/teamProject/syusyu/service/BoardServiceImpl.java
@@ -1,7 +1,8 @@
 package com.teamProject.syusyu.service;
 
-import com.teamProject.syusyu.dao.*;
-import com.teamProject.syusyu.domain.*;
+import com.teamProject.syusyu.dao.BoardDao;
+import com.teamProject.syusyu.domain.BoardDto;
+import com.teamProject.syusyu.domain.SearchCondition;
 import org.springframework.beans.factory.annotation.*;
 import org.springframework.stereotype.*;
