+++ src/main/resources/mapper/member/memberMapper.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+        "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+
+<mapper namespace="com.teamProject.syusyu.MemberMapper">
+    <!--
+    로그인 체크를 한다.
+    매개변수 : MemberDto
+    반화타입 : int
+     -->
+    <select id="loginCheck" parameterType="MemberDto" resultType="int">
+        SELECT COUNT(*)
+        FROM MEMBER
+        WHERE LGIN_ID = #{lginId}
+          AND LGIN_PWD = #{lginPwd};
+    </select>
+</mapper>
