+++ src/main/webapp/WEB-INF/spring/root-context.xml
@@ -14,14 +14,14 @@
         <property name="url" value="jdbc:log4jdbc:mysql://172.30.1.57:3306/syusyu?useUnicode=true&amp;characterEncoding=utf8" />
 <!--        <property name="driverClassName" value="com.mysql.jdbc.Driver"></property>-->
 <!--        <property name="url" value="jdbc:mysql://localhost:3306/springbasic?useUnicode=true&amp;characterEncoding=utf8"></property>-->
-        <property name="username" value="syur1"></property>
-        <property name="password" value="1234"></property>
+        <property name="username" value="syur1" />
+        <property name="password" value="1234" />
     </bean>
 
     <bean id="sqlSessionFactory" class="org.mybatis.spring.SqlSessionFactoryBean">
         <property name="dataSource" ref="dataSource"/>
         <property name="configLocation" value="classpath:mybatis-config.xml"/> <!-- mybatis 설정파일 -->
-        <property name="mapperLocations" value="classpath:mapper/*Mapper.xml"/>
+        <property name="mapperLocations" value="classpath:mapper/*/*Mapper.xml"/>
     </bean>
 
     <bean id="sqlSession" class="org.mybatis.spring.SqlSessionTemplate">
