+++ src/main/webapp/resources/static/fos/css/common/common.css
