+++ src/main/webapp/resources/static/fos/css/common/main.css
