+++ src/main/webapp/resources/static/fos/css/member/register.css
@@ -1,0 +1,13 @@
+.join_area {
+    margin: 0 auto;
+    padding: 58px 0 160px;
+    width: 400px
+}
+
+.join_title {
+    padding-bottom: 46px;
+    text-align: center;
+    font-size: 32px;
+    letter-spacing: -.48px;
+    color: #000
+}
