+++ src/main/webapp/resources/static/fos/js/index.js
